Revert "responsif_1.0"
