--- lib/colors/colors.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MyColor {
-  // --- Primary ---
-
-  // --- Seconday ---
-
-  // --- Accent ---
-}
