+++ lib/controllers/auth_controller.dart
@@ -11,7 +11,7 @@ class AuthController extends GetxController {
     String usernameToString = username.text.trim();
     String passwordToString = password.text.trim();
 
-    if (usernameToString == "madi" && passwordToString == "123") {
+    if (usernameToString == "madi" && passwordToString == "keiko") {
       Get.snackbar(
         "Auth",
         "Login Berhasil",
