--- lib/controllers/responsif_controller.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-
-class ResponsifController extends GetxController {
-  var isMobile = true.obs;
-
-  void updateScreen(BoxConstraints constraints) {
-    isMobile.value = constraints.maxWidth < 600;
-  }
-}
