--- lib/controllers/splashscreen_controller.dart
@@ -1,16 +1,0 @@
-import 'package:get/get.dart';
-import 'package:plany/routes/routes.dart';
-
-class SplashController extends GetxController {
-  @override
-  void onInit() {
-    super.onInit();
-    _navigateToLogin();
-  }
-
-  void _navigateToLogin() {
-    Future.delayed(const Duration(seconds: 3), () {
-      Get.offAllNamed(AppRoutes.login);
-    });
-  }
-}
