+++ lib/controllers/task_controller.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:plany/models/model_todo.dart';
 import 'package:plany/routes/routes.dart';
-
 class TaskController extends GetxController {
   // textfield controller
   final inputTitle = TextEditingController();
@@ -13,7 +12,6 @@ class TaskController extends GetxController {
   var activeTasks = <ModelTodo>[].obs;
   var completedTasks = <ModelTodo>[].obs;
 
-  static const List<String> categories = ['Urgent', 'Normal', 'Low'];
   // tambah todo
   void createTask() {
     final title = inputTitle.text.trim();
@@ -46,28 +44,8 @@ class TaskController extends GetxController {
 
   // hapus dari history
   void deleteHistory(int index) {
-    Get.dialog(
-      AlertDialog(
-        title: const Text("Hapus Riwayat?"),
-        content: const Text("Apakah kamu yakin ingin menghapus task ini?"),
-        actions: [
-          TextButton(
-            onPressed: () {
-              Get.back(); // tutup dialog
-            },
-            child: const Text("Tidak"),
-          ),
-          TextButton(
-            onPressed: () {
-              Get.offAllNamed(AppRoutes.history); // tutup dialog dulu
-              deleteHistory(index); // lalu hapus data
-            },
-            child: const Text("Ya"),
-          ),
-        ],
-      ),
-      barrierDismissible: false, // biar ga ketutup kalau klik di luar
-    );
+    completedTasks.removeAt(index);
+    _showMessage("Riwayat task dihapus", Colors.orangeAccent);
   }
 
   // hapus dari todo aktif
