--- lib/pages/mobileview/mobile_splash_screen.dart
@@ -1,45 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:plany/routes/routes.dart';
-
-class MobileSplashScreen extends StatefulWidget {
-  const MobileSplashScreen({super.key});
-
-  @override
-  State<MobileSplashScreen> createState() => _MobileSplashScreenState();
-}
-
-class _MobileSplashScreenState extends State<MobileSplashScreen> {
-  @override
-  void initState() {
-    super.initState();
-    Future.delayed(const Duration(seconds: 3), () {
-      Get.offAllNamed(AppRoutes.login);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: const Color(0xFF79867F),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset("assets/images/logo.png", height: 120),
-            const SizedBox(height: 16),
-            const Text(
-              "PLANY",
-              style: TextStyle(
-                fontSize: 28,
-                fontWeight: FontWeight.bold,
-                color: Colors.white,
-                letterSpacing: 2,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
