+++ lib/pages/profile_page.dart
@@ -1,28 +1,72 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:plany/controllers/responsif_controller.dart';
-import 'package:plany/pages/mobileview/mobile_profile_page.dart';
-import 'package:plany/pages/wideview/wide_profile_page.dart';
 
 class ProfilePage extends StatelessWidget {
-  ProfilePage({super.key});
-
-  final ResponsifController responsifController =
-      Get.find<ResponsifController>();
+  const ProfilePage({super.key});
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: LayoutBuilder(
-        builder: (context, constraints) {
-          responsifController.updateScreen(constraints);
+      backgroundColor: const Color(0xFF79867F),
+      appBar: AppBar(
+        backgroundColor: const Color(0xFFF1F0E4),
+        title: const Text(
+          "PROFILE",
+          style: TextStyle(fontWeight: FontWeight.bold, color: Colors.black),
+        ),
+        centerTitle: false,
+        elevation: 0,
+      ),
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const SizedBox(height: 12),
 
-          return Obx(
-            () => responsifController.isMobile.value
-                ? MobileProfilePage()
-                : WideProfilePage(),
-          );
-        },
+          // List Profile
+          Expanded(
+            child: ListView(
+              padding: const EdgeInsets.all(12),
+              children: const [
+                ProfileCard(name: "Jumadi", age: 18),
+                ProfileCard(name: "Keiko Shafira Wiyana", age: 19),
+              ],
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class ProfileCard extends StatelessWidget {
+  final String name;
+  final int age;
+
+  const ProfileCard({super.key, required this.name, required this.age});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      margin: const EdgeInsets.only(bottom: 16),
+      padding: const EdgeInsets.all(16),
+      decoration: BoxDecoration(
+        color: const Color(0xFFEDECDD), // putih krem
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const CircleAvatar(
+            radius: 40,
+            backgroundColor: Color(0xFFD9D9D9), // abu muda
+            child: Icon(Icons.person, size: 50, color: Color(0xFF3B3B2F)),
+          ),
+          const SizedBox(height: 12),
+          Text(
+            name,
+            style: const TextStyle(fontSize: 18, fontWeight: FontWeight.w500),
+          ),
+          Text("$age", style: const TextStyle(fontSize: 16)),
+        ],
       ),
     );
   }
