+++ lib/pages/splash_screen.dart
@@ -1,11 +1,22 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:plany/controllers/splashscreen_controller.dart';
+import 'package:plany/routes/routes.dart';
 
-class SplashScreen extends StatelessWidget {
-  SplashScreen({super.key});
+class SplashScreen extends StatefulWidget {
+  const SplashScreen({super.key});
 
-  final SplashController splashController = Get.put(SplashController());
+  @override
+  State<SplashScreen> createState() => _SplashScreenState();
+}
+
+class _SplashScreenState extends State<SplashScreen> {
+  @override
+  void initState() {
+    super.initState();
+    Future.delayed(const Duration(seconds: 3), () {
+      Get.offAllNamed(AppRoutes.login);
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -15,7 +26,7 @@ class SplashScreen extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            Image.asset("assets/images/logo.png", height: 200),
+            Image.asset("assets/images/logo.png", height: 120),
             const SizedBox(height: 16),
             const Text(
               "PLANY",
