+++ lib/routes/routes.dart
@@ -1,6 +1,4 @@
 import 'package:get/get.dart';
-import 'package:plany/bindings/bindings.dart';
-import 'package:plany/bindings/splash_binding.dart';
 import '../bindings/todo_binding.dart';
 import '../pages/add_todo_page.dart';
 import '../pages/history_page.dart';
@@ -25,29 +23,29 @@ class AppPages {
     GetPage(
       name: AppRoutes.splash,
       page: () => SplashScreen(),
-      binding: Binding(),
+      binding: TodoBinding(),
     ),
     GetPage(
       name: AppRoutes.initial,
       page: () => InitialPage(),
-      binding: Binding(), // tambahkan ini
+      binding: TodoBinding(), // tambahkan ini
     ),
     GetPage(name: AppRoutes.login, page: () => LoginPage()),
-    GetPage(name: AppRoutes.home, page: () => HomePage(), binding: Binding()),
+    GetPage(
+      name: AppRoutes.home,
+      page: () => HomePage(),
+      binding: TodoBinding(),
+    ),
     GetPage(
       name: AppRoutes.addTodo,
       page: () => AddTodoPage(),
-      binding: Binding(),
+      binding: TodoBinding(),
     ),
     GetPage(
       name: AppRoutes.history,
       page: () => HistoryPage(),
-      binding: Binding(),
+      binding: TodoBinding(),
     ),
-    GetPage(
-      name: AppRoutes.profile,
-      page: () => ProfilePage(),
-      binding: Binding(),
-    ),
+    GetPage(name: AppRoutes.profile, page: () => ProfilePage()),
   ];
 }
